+++ include/pajlada/settings/settingmanager.hpp
@@ -51,8 +51,8 @@ public:
         Skipped,
     };
 
-    rapidjson::Value *get(const char *path);
-    bool set(const char *path, const rapidjson::Value &value,
+    rapidjson::Value *get(const std::string &path);
+    bool set(const std::string &path, const rapidjson::Value &value,
              SignalArgs args = SignalArgs());
 
 private: